+++ app/_model/Season.ts
@@ -8,14 +8,19 @@ export interface GrandPrix {
   EventFormat: string;
   Session1: string;
   Session1Date: Date;
+  Session1DateUtc: Date;
   Session2: string;
   Session2Date: Date;
+  Session2DateUtc: Date;
   Session3: string;
   Session3Date: Date;
+  Session3DateUtc: Date;
   Session4: string;
   Session4Date: Date;
+  Session4DateUtc: Date;
   Session5: string;
   Session5Date: Date;
+  Session5DateUtc: Date;
   F1ApiSupport: boolean;
   qualifyingResults?: QualifyingResult[];
   sprintResults?: SprintResult[];
